requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pandas
matplotlib
python-docx
pypdf
requests